Avoid using -1u constant and fix conflict with max() macro from Windows API

+++ include/LIEF/PE/ExportEntry.hpp
@@ -70,7 +70,7 @@ class LIEF_API ExportEntry : public LIEF::Symbol {
   }
 
   inline void value(uint64_t value) override {
-    address(value);
+    address(static_cast<uint32_t>(value));
   }
 
   inline void set_forward_info(std::string lib, std::string function)  {
